requirements.txt
# === Core Framework ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# === Data / ML ===
numpy>=1.24.0
scikit-learn>=1.3.0
hdbscan>=0.8.33

# === Image Processing ===
Pillow>=10.0.0
opencv-python-headless>=4.8.0
imagehash>=4.3.1

# === Deep Learning ===
torch>=2.1.0
torchvision>=0.16.0
open-clip-torch>=2.23.0

# === Face Analysis ===
insightface>=0.7.3
onnxruntime>=1.16.0

# === Utilities ===
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
tqdm>=4.66.0
